node: documentation for add, answer, and vote code
Questions:

contributor/index.js, line 30
I don't see patchVoteCount() called anywhere in /lib. Can we delete?

qnaDoc/patch.js, line 132
The spec param is passed in with an empty comments property. Is this needed?

qnaDoc/index.js, line 31
The spec param can specify a single document or multiple (as coded). Where is an example of that method handling multiple documents? I don't see any examples. In such a case, is spec an array of objects (similar to the object for the singular case)?